read_image arguments can now be char *s
git-svn-id: https://serveur-svn.lri.fr/svn/modhel/luatex/trunk@2569 0b2b3880-5936-4365-a048-eb17d2e5a6bf

+++ source/texk/web2c/luatexdir/pdf/pdfshipout.c
@@ -793,7 +793,7 @@ void pdf_print_info(integer luatex_version, str_number luatex_revision)
     boolean creator_given, producer_given, creationdate_given, moddate_given,
         trapped_given;
     char *s = NULL;
-    integer len = 0;
+    int len = 0;
     pdf_new_dict(obj_type_others, 0, 3);        /* keep Info readable unless explicitely forced */
     creator_given = false;
     producer_given = false;
@@ -1280,18 +1280,18 @@ void finish_pdf_file(integer luatex_version, str_number luatex_revision)
         if (log_opened) {
             fprintf(log_file,
                     "\nPDF statistics: %d PDF objects out of %d (max. %d)\n",
-                    obj_ptr, obj_tab_size, sup_obj_tab_size);
+                    (int)obj_ptr, (int)obj_tab_size, (int)sup_obj_tab_size);
             if (pdf_os_cntr > 0) {
                 fprintf(log_file,
                         " %d compressed objects within %d object stream%s\n",
-                        ((pdf_os_cntr - 1) * pdf_os_max_objs + pdf_os_objidx +
-                         1), pdf_os_cntr, (pdf_os_cntr > 1 ? "s" : ""));
+                        (int)((pdf_os_cntr - 1) * pdf_os_max_objs + pdf_os_objidx +
+                         1), (int)pdf_os_cntr, (pdf_os_cntr > 1 ? "s" : ""));
             }
             fprintf(log_file, " %d named destinations out of %d (max. %d)\n",
-                    pdf_dest_names_ptr, dest_names_size, sup_dest_names_size);
+                    (int)pdf_dest_names_ptr, (int)dest_names_size, (int)sup_dest_names_size);
             fprintf(log_file,
                     " %d words of extra memory for PDF output out of %d (max. %d)\n",
-                    pdf_mem_ptr, pdf_mem_size, sup_pdf_mem_size);
+                    (int)pdf_mem_ptr, (int)pdf_mem_size, (int)sup_pdf_mem_size);
         }
     }
 }
